// src/app/features/games/slotMachine/enums/slotVariants.enum.ts
export enum SlotVariants {
    // TOTAL WIN
    AXE = "AXE",
    GOLD = "GOLD",
    CHERRY = "CHERRY",
    FRUITS = "FRUITS",
    SLOT_WIN = "SLOT_WIN",
    // CLOSE TO WIN
    CLOSE_AXE = "CLOSE_AXE",
    CLOSE_GOLD = "CLOSE_GOLD",
    CLOSE_CHERRY = "CLOSE_CHERRY",
    CLOSE_FRUITS = "CLOSE_FRUITS",
    CLOSE_SLOT_WIN = "CLOSE_SLOT_WIN",

    // LOSE
    LOSE = "LOSE",
}